bt-172: * fixed paths
