+++ backend/src/bundles/user-details/enums/enums.ts
@@ -1,0 +1,9 @@
+export {
+    CountryList,
+    EmploymentType,
+    EnglishLevel,
+    JobTitle,
+    NotConsidered,
+    PreferredLanguages,
+    UserDetailsApiPath,
+} from 'shared/build/index.js';
