+++ backend/src/bundles/user-details/types/types.ts
@@ -1,1 +1,6 @@
 export { type UserDetailsProperties } from './user-details-properties.type.js';
+export {
+    type UserDetailsApproveRequestDto,
+    type UserDetailsResponseDto,
+    type UserDetailsUpdateRequestDto,
+} from 'shared/build/index.js';
