+++ backend/src/bundles/user-details/validation-schemas/validation-schemas.ts
@@ -1,0 +1,4 @@
+export {
+    userDetailsApproveValidationSchema,
+    userDetailsUpdateValidationSchema,
+} from 'shared/build/index.js';
