+++ backend/src/bundles/users/enums/enums.ts
@@ -1,9 +1,1 @@
-export {
-    CountryList,
-    EmploymentType,
-    EnglishLevel,
-    JobTitle,
-    NotConsidered,
-    PreferredLanguages,
-    UsersApiPath,
-} from 'shared/build/index.js';
+export { UsersApiPath } from 'shared/build/index.js';
