+++ shared/src/bundles/user-details/enums/country-list.enum.ts
