+++ shared/src/bundles/user-details/enums/employment-type.enum.ts
