+++ shared/src/bundles/user-details/enums/job-title.enum.ts
