+++ shared/src/bundles/user-details/enums/not-considered.enum.ts
