+++ shared/src/bundles/user-details/enums/user-details-api-path.enum.ts
