+++ shared/src/bundles/user-details/enums/user-role.enum.ts
@@ -1,0 +1,7 @@
+const UserRole = {
+    TALENT: 'talent',
+    EMPLOYER: 'employer',
+    ADMIN: 'admin',
+} as const;
+
+export { UserRole };
