+++ shared/src/bundles/user-details/enums/user-validation-message.enum.ts
@@ -1,0 +1,6 @@
+const UserValidationMessage = {
+    EMAIL_REQUIRE: 'Email is required',
+    EMAIL_WRONG: 'Email is wrong',
+} as const;
+
+export { UserValidationMessage };
