+++ shared/src/bundles/user-details/types/types.ts
@@ -1,0 +1,3 @@
+export { type UserDetailsApproveRequestDto } from './user-details-approve-request-dto.type.js';
+export { type UserDetailsResponseDto } from './user-details-response-dto.type.js';
+export { type UserDetailsUpdateRequestDto } from './user-details-update-request-dto.type.js';
