+++ shared/src/bundles/user-details/types/user-details-approve-request-dto.type.ts
