+++ shared/src/bundles/user-details/types/user-details-update-request-dto.type.ts
@@ -7,7 +7,7 @@ import {
     type JobTitle,
     type NotConsidered,
     type PreferredLanguages,
-} from '../users.js';
+} from '../user-details.js';
 
 type UserDetailsUpdateRequestDto = {
     id: string;
