+++ shared/src/bundles/user-details/user-details.ts
@@ -1,0 +1,16 @@
+export {
+    CountryList,
+    EmploymentType,
+    EnglishLevel,
+    JobTitle,
+    NotConsidered,
+    PreferredLanguages,
+    UserDetailsApiPath,
+} from './enums/enums.js';
+export {
+    type UserDetailsApproveRequestDto,
+    type UserDetailsResponseDto,
+    type UserDetailsUpdateRequestDto,
+} from './types/types.js';
+export { userDetailsUpdate as userDetailsUpdateValidationSchema } from './validation-schemas/validation-schemas.js';
+export { userDetailsApprove as userDetailsApproveValidationSchema } from './validation-schemas/validation-schemas.js';
