+++ shared/src/bundles/user-details/validation-schemas/validation-schemas.ts
@@ -1,0 +1,2 @@
+export { userDetailsApprove } from './user-details-approve.validation-schema.js';
+export { userDetailsUpdate } from './user-details-update.validation-schema.js';
