+++ shared/src/bundles/users/enums/enums.ts
@@ -1,10 +1,3 @@
-export { CountryList } from './country-list.enum.js';
-export { EmploymentType } from './employment-type.enum.js';
-export { EnglishLevel } from './english-level.enum.js';
-export { JobTitle } from './job-title.enum.js';
-export { NotConsidered } from './not-considered.enum.js';
-export { PreferredLanguages } from './preferred-languages.enum.js';
-export { UserDetailsApiPath } from './user-details-api-path.enum.js';
 export { UserRole } from './user-role.enum.js';
 export { UserValidationMessage } from './user-validation-message.enum.js';
 export { UsersApiPath } from './users-api-path.enum.js';
