+++ shared/src/bundles/users/types/types.ts
@@ -1,7 +1,4 @@
 export { type UserCreateResponseDto } from './user-create-response-dto.type.js';
-export { type UserDetailsApproveRequestDto } from './user-details-approve-request-dto.type.js';
-export { type UserDetailsResponseDto } from './user-details-response-dto.type.js';
-export { type UserDetailsUpdateRequestDto } from './user-details-update-request-dto.type.js';
 export { type UserFindResponseDto } from './user-find-response-dto.type.js';
 export { type UserGetAllItemResponseDto } from './user-get-all-item-response-dto.type.js';
 export { type UserGetAllResponseDto } from './user-get-all-response-dto.type.js';
