+++ shared/src/bundles/users/users.ts
@@ -1,20 +1,10 @@
 export {
-    CountryList,
-    EmploymentType,
-    EnglishLevel,
-    JobTitle,
-    NotConsidered,
-    PreferredLanguages,
-    UserDetailsApiPath,
     UserRole,
     UsersApiPath,
     UserValidationMessage,
 } from './enums/enums.js';
 export {
     type UserCreateResponseDto,
-    type UserDetailsApproveRequestDto,
-    type UserDetailsResponseDto,
-    type UserDetailsUpdateRequestDto,
     type UserFindResponseDto,
     type UserGetAllItemResponseDto,
     type UserGetAllResponseDto,
@@ -25,5 +15,3 @@ export {
 } from './types/types.js';
 export { userSignUp as userSignUpValidationSchema } from './validation-schemas/validation-schemas.js';
 export { userSignIn as userSignInValidationSchema } from './validation-schemas/validation-schemas.js';
-export { userDetailsUpdate as userDetailsUpdateValidationSchema } from './validation-schemas/validation-schemas.js';
-export { userDetailsApprove as userDetailsApproveValidationSchema } from './validation-schemas/validation-schemas.js';
