+++ shared/src/bundles/users/validation-schemas/validation-schemas.ts
@@ -1,4 +1,2 @@
-export { userDetailsApprove } from './user-details-approve.validation-schema copy.js';
-export { userDetailsUpdate } from './user-details-update.validation-schema.js';
 export { userSignIn } from './user-sign-in.validation-schema.js';
 export { userSignUp } from './user-sign-up.validation-schema.js';
